+++ squadcastv1/src/squadcast/models/schedules_listschedulesop.py
@@ -139,7 +139,7 @@ class SchedulesListSchedulesResponseBodyTypedDict(TypedDict):
     r"""The request has succeeded."""
 
     data: List[V4ScheduleResponseTypedDict]
-    page_info: CommonV4PageInfoTypedDict
+    page_info: NotRequired[CommonV4PageInfoTypedDict]
 
 
 class SchedulesListSchedulesResponseBody(BaseModel):
@@ -147,7 +147,25 @@ class SchedulesListSchedulesResponseBody(BaseModel):
 
     data: List[V4ScheduleResponse]
 
-    page_info: Annotated[CommonV4PageInfo, pydantic.Field(alias="pageInfo")]
+    page_info: Annotated[
+        Optional[CommonV4PageInfo], pydantic.Field(alias="pageInfo")
+    ] = None
+
+    @model_serializer(mode="wrap")
+    def serialize_model(self, handler):
+        optional_fields = set(["pageInfo"])
+        serialized = handler(self)
+        m = {}
+
+        for n, f in type(self).model_fields.items():
+            k = f.alias or n
+            val = serialized.get(k, serialized.get(n))
+
+            if val != UNSET_SENTINEL:
+                if val is not None or k not in optional_fields:
+                    m[k] = val
+
+        return m
 
 
 class SchedulesListSchedulesResponseTypedDict(TypedDict):